requirements-test.txt
torch>=2.4.1
torchaudio>=2.4.1
soundfile>=0.12.1
librosa>=0.10.1
numpy
scipy
julius
omegaconf>=2.3.0
pytest
faster-whisper>=0.9.0
huggingface-hub>=0.23.4
wavmark>=0.0.3
pydub>=0.25.1
pyyaml>=6.0
ffmpeg-python>=0.2.0
gdown>=5.2.0
pillow>=8.4.0
matplotlib
seaborn
numba>=0.54.1
pandas
scikit-learn
tqdm
pyaml>=24.4.0
build
Flask
black
isort
flake8
pre-commit>=3
cython>=3.0.0
eng_to_ipa>=0.0.2
inflect>=7.0.0
unidecode>=1.3.7
whisper-timestamped>=1.14.2
pypinyin>=0.50.0
cn2an>=0.5.22
jieba>=0.42.1
gradio>=3.48.0
langid>=1.1.6
openai-whisper
encodec>=0.1.1
sudachipy
sudachidict_core
vocos
nltk
packaging>=23.1
coqui-tts-trainer>=0.1.4
coqpit>=0.0.16
gruut[de,es,fr]>=2.4.0
anyascii>=0.3.0
transformers>=4.43.0
spacy[ja]>=3,<3.8
boto3
funcy
wheel
wandb
wget
tensorboardx
setuptools>=18.0
einops>=0.6.0
tokenizers
audiolm-pytorch
-e ./audioseal
-e ./OpenVoice
-e ./coqui-ai-TTS
-e ./fairseq
#-e ./bark-with-voice-clone